+++ administration/src/application/components/forms/ApplicationForm.tsx
@@ -1,11 +1,21 @@
 import { SetState, useUpdateStateCallback } from './useUpdateStateCallback'
-import { BlueCardEntitlementType } from '../../../generated/graphql'
 import {
+  ApplicationType,
+  BlueCardApplicationInput,
+  BlueCardEntitlementType,
+} from '../../../generated/graphql'
+import {
+  convertStandardEntitlementFormStateToInput,
   initialStandardEntitlementFormState,
   StandardEntitlementForm,
   StandardEntitlementFormState,
 } from './StandardEntitlementForm'
-import { initialPersonalDataFormState, PersonalDataForm, PersonalDataFormState } from './PersonalDataForm'
+import {
+  convertPersonalDataFormStateToInput,
+  initialPersonalDataFormState,
+  PersonalDataForm,
+  PersonalDataFormState,
+} from './PersonalDataForm'
 import SwitchDisplay from '../SwitchDisplay'
 import { FormControl, FormControlLabel, FormLabel, Radio, RadioGroup } from '@mui/material'
 
@@ -63,8 +73,8 @@ export const ApplicationForm = ({
               setState={useUpdateStateCallback(setState, 'standardEntitlement')}
             />
           ),
-          [BlueCardEntitlementType.Juleica]: <JuleicaEntitlementForm />,
-          [BlueCardEntitlementType.Service]: <ServiceEntitlementForm />,
+          [BlueCardEntitlementType.Juleica]: null,
+          [BlueCardEntitlementType.Service]: null,
         }}
       </SwitchDisplay>
       <PersonalDataForm state={state.personalData} setState={useUpdateStateCallback(setState, 'personalData')} />
@@ -72,5 +82,26 @@ export const ApplicationForm = ({
   )
 }
 
-const JuleicaEntitlementForm = () => null
-const ServiceEntitlementForm = () => null
+export const convertApplicationFormStateToInput = (state: ApplicationFormState): BlueCardApplicationInput => {
+  const entitlement = (() => {
+    if (state.entitlementType === null) throw Error('EntitlementType is null.')
+    switch (state.entitlementType) {
+      case BlueCardEntitlementType.Standard:
+        const workAtOrganizations = convertStandardEntitlementFormStateToInput(state.standardEntitlement)
+        return {
+          entitlementType: state.entitlementType,
+          workAtOrganizations,
+        }
+      default:
+        throw Error('Not yet implemented.')
+    }
+  })()
+
+  return {
+    entitlement,
+    personalData: convertPersonalDataFormStateToInput(state.personalData),
+    hasAcceptedPrivacyPolicy: true, // TODO: Add a corresponding field
+    applicationType: ApplicationType.FirstApplication, // TODO: Add a corresponding field
+    givenInformationIsCorrectAndComplete: true, // TODO: Add a corresponding field
+  }
+}
